test(python-sdk): move the registry unit tests out of the e2e tree
The tests are in-memory and need no backend, so tests/e2e/ was the wrong home
for them and they carried no marker. Moves them to tests/utils/ beside the
suite's other unit tests, marks them unit, and moves the registry itself into
tests/fixtures/, which is already the shared test-support package the rest of
the suite imports as a top-level module.
