+++ sdks/python/tests/fixtures/tag_registry.py
